+++ hydromt/metadata_resolver/convention_resolver.py
@@ -7,11 +7,10 @@ from re import error as regex_error
 from string import Formatter
 from typing import Any, List, Optional, Pattern, Set, Tuple
 
-import geopandas as gpd
 import numpy as np
 import pandas as pd
 
-from hydromt._typing import Bbox, NoDataStrategy, Predicate, TimeRange
+from hydromt._typing import Bbox, Geom, NoDataStrategy, Predicate, TimeRange
 
 from .metadata_resolver import MetaDataResolver
 
@@ -89,8 +88,8 @@ class ConventionResolver(MetaDataResolver):
         *,
         timerange: Optional[TimeRange] = None,
         bbox: Optional[Bbox] = None,
-        # TODO: align? -> from RasterDatasetAdapter
-        mask: Optional[gpd.GeoDataFrame] = None,
+        # TODO: align? https://github.com/Deltares/hydromt/issues/874
+        mask: Optional[Geom] = None,
         buffer: float = 0.0,
         predicate: Predicate = "intersects",
         variables: Optional[List[str]] = None,